+++ run.sh
@@ -1,6 +1,6 @@
 export OMP_NUM_THREADS=1
-export CUDA_VISIBLE_DEVICES=0,1,2,3
+export CUDA_VISIBLE_DEVICES=4,5,6,7
 torchrun \
   --nproc_per_node 4 \
   --master_port 29550 \
-  train_class_GCM.py
+  train_classify_seg_GCNC.py
